+++ src/packages/relay/src/createFragmentContainer.js
@@ -4,7 +4,8 @@ import * as React from 'react';
 import { createFragmentContainer as _createFragmentContainer } from 'react-relay';
 import { invariant, isObjectEmpty } from '@kiwicom/js';
 
-import type { FragmentSpec, $RelayProps, Environment } from './types.flow';
+import type { FragmentSpec, $RelayProps } from './types.flow';
+import type { Environment } from './runtimeTypes.flow';
 
 export type RelayProp = {|
   +environment: Environment,