+++ src/packages/relay/src/createRefetchContainer.js
@@ -4,13 +4,8 @@ import * as React from 'react';
 import { createRefetchContainer as _createRefetchContainer } from 'react-relay';
 import { invariant, isObjectEmpty } from '@kiwicom/js';
 
-import type {
-  $RelayProps,
-  FragmentSpec,
-  GraphQLTaggedNode,
-  Disposable,
-  Environment,
-} from './types.flow';
+import type { $RelayProps, FragmentSpec, GraphQLTaggedNode, Disposable } from './types.flow';
+import type { Environment } from './runtimeTypes.flow';
 
 type RefetchOptions = { force?: boolean, ... };