+++ src/packages/relay/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Improved flow type coverage of Relay store
+
 # 4.6.0
 
 - Fixed Flow types for `createFragmentContainer`, `createPaginationContainer` and `createRefetchContainer`. This may yield many new errors especially if you didn't type your React components correctly. But don't worry, the upgrade can be very simple and automated - contact us directly. :)